requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
py-trello==0.14.0
pylint==2.2.2
python-dateutil==2.7.5
python-dotenv==0.10.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
typed-ast==1.1.0
urllib3==1.24.1
wrapt==1.10.11
